requirements.txt
absl-py==2.4.0
altgraph==0.17.5
cffi==2.0.0
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.62.1
kiwisolver==1.5.0
matplotlib==3.10.8
mediapipe==0.10.33
numpy==2.4.3
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
packaging==26.0
pefile==2024.8.26
pillow==12.1.1
pycparser==3.0
pygame-ce==2.5.7
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
setuptools==82.0.1
six==1.17.0
sounddevice==0.5.5
cv2-enumerate-cameras==1.3.3
colorama==0.4.6
coverage==7.13.5
iniconfig==2.3.0
pluggy==1.6.0
pygments==2.19.2
pytest==9.0.2
pytest-cov==7.1.0
